NXBT-418: Add EAR test script - work in progress

+++ integration-lib.sh
@@ -129,13 +129,10 @@ check_ports_and_kill_ghost_process() {
 # IT tests should not call that function, but download sources to be tested
 update_distribution_source() {
     if [ ! -d "$NXDISTRIBUTION" ]; then
-        [ ! -d `dirname "$NXDISTRIBUTION"` ] && mkdir -p `dirname "$NXDISTRIBUTION"`
+        mkdir -p `dirname "$NXDISTRIBUTION"`
         git clone git@github.com:nuxeo/nuxeo-distribution.git "$NXDISTRIBUTION" || exit 1
-        # TODO: support NXVERSION checkout
-    else
-        (cd "$NXDISTRIBUTION" && git checkout master && git pull) || exit 1
-        # TODO: support NXVERSION checkout
     fi
+    (cd "$NXDISTRIBUTION" && git checkout master && git fetch --tags && git pull --all&& git checkout $NXVERSION) || exit 1
     if [ ! -z $NXTAG ]; then
         (cd "$NXDISTRIBUTION" && git checkout $NXTAG) || exit 1
     fi
@@ -190,7 +187,9 @@ EOF
     if [ "$SERVER" = jboss ]; then
         set_jboss_log4j_level $SERVER_HOME INFO
         echo "org.nuxeo.systemlog.token=dolog" > "$SERVER_HOME"/templates/common/config/selenium.properties
-        cp "$SERVER_HOME"/server/default/data/NXRuntime/data/installAfterRestart-DM.log "$SERVER_HOME"/server/default/data/NXRuntime/data/installAfterRestart.log
+        if [ -f "$SERVER_HOME"/server/default/data/NXRuntime/data/installAfterRestart-DM.log ]; then
+            cp "$SERVER_HOME"/server/default/data/NXRuntime/data/installAfterRestart-DM.log "$SERVER_HOME"/server/default/data/NXRuntime/data/installAfterRestart.log
+        fi
         mkdir -p "$SERVER_HOME"/log
     fi
     if [ "$SERVER" = tomcat ]; then
@@ -212,12 +211,14 @@ setup_jboss() {
         cp -r "$NXDISTRIBUTION"/nuxeo-distribution-jboss/target/nuxeo-*-jboss "$JBOSS" || exit 1
     else
         echo "Using previously installed JBoss. Set NEW_JBOSS variable to force new JBOSS deployment"
-        mkdir -p saved_files
-        mv "$JBOSS"/server/default/data/NXRuntime/data/installAfterRestart* saved_files/
-        rm -rf "$JBOSS"/server/default/data/* "$JBOSS"/log/*
-        mkdir -p "$JBOSS"/server/default/data/NXRuntime/data
-        mv saved_files/* "$JBOSS"/server/default/data/NXRuntime/data/
-        rmdir saved_files
+        if [ -f "$SERVER_HOME"/server/default/data/NXRuntime/data/installAfterRestart-DM.log ]; then
+            mkdir -p saved_files
+            mv "$JBOSS"/server/default/data/NXRuntime/data/installAfterRestart* saved_files/
+            rm -rf "$JBOSS"/server/default/data/* "$JBOSS"/log/*
+            mkdir -p "$JBOSS"/server/default/data/NXRuntime/data
+            mv saved_files/* "$JBOSS"/server/default/data/NXRuntime/data/
+            rmdir saved_files
+        fi
     fi
     mkdir -p "$JBOSS"/log
     setup_server_conf $JBOSS
